Build: Remove curl write callback compiler warning

+++ roam/roam_check_access.c
@@ -155,7 +155,7 @@ struct _buf {
   char *ptr;
 };
 
-static int callback(char *ptr, size_t size, size_t nmemb, struct _buf *buf)
+static size_t callback(char *ptr, size_t size, size_t nmemb, struct _buf *buf)
 {
   if (buf->size == 0) {
     buf->ptr = malloc(size * nmemb + 1);
